+++ src/main/java/com/mojang/minecraft/renderer/graphics/opengl/OpenGLIndexBuffer.java
@@ -91,7 +91,7 @@ public class OpenGLIndexBuffer implements IndexBuffer {
     }
     
     @Override
-    public int getSizeInBytes() {
+    public long getSizeInBytes() {
         return sizeInBytes;
     }
